Reorganize accounting unit tests into a separate directory
